FIx pathing issues and mult-downloads
Added feature to move all media in a playlist to a different directory
as needed. Same for a file with fullpath that is DNE and move the file
to that new directory.

Added start and stop positions in the playlist for testing purpose in
order to manage time better when checking for media names that are
downloaded

Added timeout and keyboard interrupt

Moved the model_bin check to the beginning